Se corrige el error de los campos course_id y subject_id de la tabla course_subject para la relación, Se creo un seeder subject para insertar datos específicos de las materias,  Se creo un factory student para generar datos aleotorios de los estudiantes y  Se creo la api básico de la tabla Student

+++ database/migrations/2024_05_29_163515_create_course_subject_table.php
@@ -14,10 +14,10 @@ return new class extends Migration
         Schema::create('course_subject', function (Blueprint $table) {
             $table-> id ();
 
-            $table-> unsignedInteger ('course_id');
-            $table-> unsignedInteger ('subject_id');
+            $table-> unsignedBigInteger ('course_id');
+            $table-> unsignedBigInteger ('subject_id');
 
-            $table->foreign ('course_id')-> references ('id')-> on ('courses')-> onDelete ('cascade');
+            $table-> foreign ('course_id')-> references('id')-> on('courses')-> onDelete('cascade');
             $table->foreign ('subject_id')-> references ('id')-> on ('subjects')-> onDelete ('cascade'); 
 
             $table->timestamps();
